chore: Move virtio_gen module to vmm::devices::virtio
In our ongoing effort to eliminate our many crates, merge the virtio
bindings into the virtio modules of the vmm crate.
