--- src/virtio_gen/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "virtio_gen"
-version = "0.1.0"
-authors = ["The Chromium OS Authors"]
-edition = "2021"
-license = "Apache-2.0"
-
-[lib]
-bench = false
