+++ src/vmm/src/devices/virtio/gen/virtio_ring.rs
