+++ src/vmm/src/devices/virtio/gen/virtio_rng.rs
