+++ src/vmm/src/devices/virtio/rng/device.rs
@@ -8,10 +8,10 @@ use std::sync::Arc;
 use aws_lc_rs::rand;
 use utils::eventfd::EventFd;
 use utils::vm_memory::{GuestMemoryError, GuestMemoryMmap};
-use virtio_gen::virtio_rng::VIRTIO_F_VERSION_1;
 
 use super::{RNG_NUM_QUEUES, RNG_QUEUE};
 use crate::devices::virtio::device::{IrqTrigger, IrqType};
+use crate::devices::virtio::gen::virtio_rng::VIRTIO_F_VERSION_1;
 use crate::devices::virtio::iovec::IoVecBufferMut;
 use crate::devices::virtio::{
     ActivateError, DeviceState, Queue, VirtioDevice, FIRECRACKER_MAX_QUEUE_SIZE, TYPE_RNG,
